requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
httpx>=0.24.0
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=1.4
asyncpg>=0.28.0
